+++ spec/hashie/dash_spec.rb
@@ -192,6 +192,32 @@ describe DashTest do
     end
   end
 
+  context 'converting from a Mash' do
+    class ConvertingFromMash < Hashie::Dash
+      property :property, required: true
+    end
+
+    context 'without keeping the original keys' do
+      let(:mash) { Hashie::Mash.new(property: 'test') }
+
+      it 'does not pick up the property from the stringified key' do
+        expect { ConvertingFromMash.new(mash) }.to raise_error(NoMethodError)
+      end
+    end
+
+    context 'when keeping the original keys' do
+      class KeepingMash < Hashie::Mash
+        include Hashie::Extensions::Mash::KeepOriginalKeys
+      end
+
+      let(:mash) { KeepingMash.new(property: 'test') }
+
+      it 'picks up the property from the original key' do
+        expect { ConvertingFromMash.new(mash) }.not_to raise_error
+      end
+    end
+  end
+
   describe '#new' do
     it 'fails with non-existent properties' do
       expect { described_class.new(bork: '') }.to raise_error(*no_property_error('bork'))